Make requests to ajax-api in JavaScript relative (#1413)
Make requests to ajax-api in JavaScript relative, so the mlflow UI can now be served from a (non-root) path behind a proxy.

When running the mlflow UI behind a proxy, it may no longer be served directly
 on `/` but in a subpath. As it is a single page web-app, the JavaScript code will
always be executed in the same path and thus we don't need any absolute references.

Fixes #1120
